Add a foundation folder.  Move Layout, Toast and theme concerns into the foundation folder.  That way App stays quite lean
