+++ packages/ui-extensions-dev-console/src/App.tsx
@@ -1,14 +1,13 @@
-import * as styles from './theme.module.css'
 import React from 'react'
 import '@shopify/polaris/dist/styles.css'
 import enTranslations from '@shopify/polaris/locales/en.json'
 import {AppProvider} from '@shopify/polaris'
 import {I18nContext, I18nManager} from '@shopify/react-i18n'
 import {ExtensionServerProvider, isValidSurface} from '@shopify/ui-extensions-server-kit'
-import {ToastContainer} from 'react-toastify'
-import Layout from '@/Layout/Layout'
-import Routes from '@/Routes'
-import 'react-toastify/dist/ReactToastify.css'
+import {Layout} from '@/foundation/Layout'
+import {Routes} from '@/foundation/Routes'
+import {Toast} from '@/foundation/Toast'
+import {Theme} from '@/foundation/Theme'
 
 function getConnectionUrl() {
   if (import.meta.env.VITE_CONNECTION_URL) {
@@ -28,9 +27,8 @@ const extensionServerOptions = {
   surface: isValidSurface(surface) ? surface : undefined,
 }
 
-const locale = 'en'
 const i18nManager = new I18nManager({
-  locale,
+  locale: 'en',
   onError(error) {
     // eslint-disable-next-line no-console
     console.log(error)
@@ -39,33 +37,18 @@ const i18nManager = new I18nManager({
 
 function App() {
   return (
-    <div className={styles.Theme}>
-      <ExtensionServerProvider options={extensionServerOptions}>
-        <I18nContext.Provider value={i18nManager}>
-          <AppProvider i18n={enTranslations}>
+    <ExtensionServerProvider options={extensionServerOptions}>
+      <I18nContext.Provider value={i18nManager}>
+        <AppProvider i18n={enTranslations}>
+          <Theme>
             <Layout>
               <Routes />
-              <ToastContainer
-                position="bottom-center"
-                theme="dark"
-                closeButton={false}
-                pauseOnFocusLoss={false}
-                closeOnClick
-                hideProgressBar
-              />
-              {/* react-toastify requires exactly this class. */}
-              {/* Inlining the style prevents the className from changing */}
-              <style>
-                {`.Toastify__toast-body {
-                  padding-top: 0;
-                  padding-bottom: 0;
-                }`}
-              </style>
+              <Toast />
             </Layout>
-          </AppProvider>
-        </I18nContext.Provider>
-      </ExtensionServerProvider>
-    </div>
+          </Theme>
+        </AppProvider>
+      </I18nContext.Provider>
+    </ExtensionServerProvider>
   )
 }
