+++ packages/ui-extensions-dev-console/src/foundation/Layout/index.ts
@@ -1,0 +1,1 @@
+export * from './Layout'
