+++ packages/ui-extensions-dev-console/src/foundation/Layout/translations/en.json
