+++ packages/ui-extensions-dev-console/src/foundation/Routes/index.ts
@@ -1,0 +1,1 @@
+export * from './Routes'
