+++ packages/ui-extensions-dev-console/src/foundation/Theme/index.ts
@@ -1,0 +1,1 @@
+export * from './Theme'
